feat: add render in storybook file
