+++ src/components/select/Select.module.scss
@@ -1,6 +1,5 @@
 .container {
   // width: 100%; //оч мешает
-  display: flex; //если не добавить то контейнер меньше содержимого
 }
 
 .label {
